requirements.txt
faiss-cpu~=1.9.0.post1
numpy~=1.26.4
pandas~=2.2.3
Flask~=3.1.0
sentence-transformers~=3.3.1
SQLAlchemy~=2.0.36
openai~=1.57.4